It closes #82.
chore: add basic implementation of EventPublisher.

+++ quakeml/src/main/scala/quakeml/Quantity.scala
@@ -4,7 +4,7 @@ import java.time.ZonedDateTime
 
 case class TimeQuantity(
     value: ZonedDateTime,
-    uncertainty: Option[Int]
+    uncertainty: Option[Double]
 )
 
 case class RealQuantity(
